Extensive Rewrite
When in the bubble it uses cached data so EDSM query is kept to a minimum.
First pass uses cached data to pick a star. It then queries EDSM in a background process and updates the UI if a star is found nearer.
Star name is coloured green when final decision on the closest star is made. 
Clicking on name only returns the name